requirements.txt
python-decouple
imutils
opencv-python